% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildcvmods.R
\name{buildcvmods}
\alias{buildcvmods}
\title{Outputs a list of models built without the ith fold}
\usage{
buildcvmods(data, target, stratify, preds, modelfunc, numfolds)
}
\arguments{
\item{data}{the dataset}

\item{target}{the target variable for the model}

\item{stratify}{the name of the field in the dataset which we wish to use to stratify the cv folds}

\item{preds}{the predictors for the model}

\item{modelfunc}{a function that takes the target and dataset with only predictors and creates a model in the desired format. The user is responsible for writing this function and ensuring it outputs the model in the desired format.}

\item{numfolds}{the number of desired folds for cross validation}
}
\value{
a list of numfolds models
}
\description{
Outputs a list of models built without the ith fold
}
\examples{
mock_data <- data.frame(preds1 = rep(1:4,10),preds2 = rep(1:10,4),target=rep(1:5,8),strat=rep(1:2,20))
model function:
model_function<-function(target,data){
targnum<-which(colnames(data) == target)
moddat<-data[,-targnum]
targ<-data[,targnum]
set.seed(123)
mod<-lm(targ~.,data=moddat)
mod
}
test function:
model_function("target",mock_data)
compare to:
lm(target~.,data=mock_data)
finally, test function:
buildcvmods(mock_data,"target","strat",c("preds1","preds2"),model_function,3)
}
