% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildcvmods.R
\name{buildcvmods}
\alias{buildcvmods}
\title{Outputs a list of models built without the ith fold. If the dataset has a column named "fold" that will be used for CV, otherwise folds will be created by the function.}
\usage{
buildcvmods(data, target, stratify, preds, modelfunc, weights, numfolds)
}
\arguments{
\item{data}{the dataset}

\item{target}{the target variable for the model}

\item{stratify}{optional: the name of the field in the dataset which we wish to use to stratify the cv folds}

\item{preds}{the predictors for the model}

\item{modelfunc}{a function that takes the target and dataset with only predictors and creates a model in the desired format. The user is responsible for writing this function and ensuring it outputs the model in the desired format.}

\item{weights}{the weights to be used in model building if desired. This must be a field in the dataframe and a column of 1s can be added if no weights are desired.}

\item{numfolds}{optional: the number of desired folds for cross validation if folds are not provided}
}
\value{
a list of models, each built with one fold of data left out
}
\description{
Outputs a list of models built without the ith fold. If the dataset has a column named "fold" that will be used for CV, otherwise folds will be created by the function.
}
\examples{
mock_data <- data.frame(preds1 = rep(1:4,10),preds2 = rep(1:10,4),target=rep(1:5,8),strat=rep(1:2,20),weights = rep(c(2,1),20))
model function:
model_function<-function(target,data,weights){
targnum<-which(colnames(data) == target)
weightnum<-which(colnames(data) == weights)
moddat<-data[,-c(targnum,weightnum)]
targ<-data[,targnum]
weights<-data[,weightnum]
set.seed(123)
mod<-lm(targ~.,data=moddat,weights=weights)
mod
}
test function:
model_function("target",mock_data,"weights")
compare to:
mock_data2 <- mock_data[,-which(colnames(mock_data) == "weights")]
lm(target~.,data=mock_data2,weights=mock_data$weights)
finally, test function:
buildcvmods(mock_data,"target","strat",c("preds1","preds2"),model_function,"weights",3)
test with no stratifying argument
buildcvmods(mock_data,"target",preds=c("preds1","preds2"),modelfunc=model_function,weights="weights",numfolds=3)
test with predefined folds
mock_data2<-fold_field(mock_data,3)
buildcvmods(mock_data2,"target",preds=c("preds1","preds2"),modelfunc=model_function,weights="weights")
}
